Migration_Jira_To_Gitlab/jira2gitlab-final/requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
idna==2.10
importlib-metadata==4.6.1
jeepney==0.6.0
jira==3.0.1
jira2markdown==0.2.0
keyring==23.0.1
numpy==1.19.5
oauthlib==3.1.1
pandas==1.1.5
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-gitlab==2.7.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.5.0